Refactored SeeFloor classes: got rid of SeefloorNoBadBlocks class, replaced it with SeeFloor.
